feat: Add usage telemetry
For CodeScene to be able to improve the VS Code extension, we've added
basic telemetry. Each time the extension is activated, we'll send a
message to devtools.codescene.io.

We're using the official TelemetryLogger interface, which respects
privacy settings and provides information on the VS Code client, like:
- vs-code version
- platform version
- extension version
- vs-code machine id, generated by vs code

The interface is the same that's used by Microsoft's own plugin, see
https://github.com/microsoft/vscode-extension-telemetry/tree/main#common-properties
for more details.