Add standard-version to handle versioning
This handles versions in `package.json` (like the normal `npm version`)
but it also updates `CHANGELOG.md` and commits it in the same commit.

This changelog is different from the one created for the releases
on github. It doesn't list all commits but only the most relevant
changes to the end-user (see `.versionrc.json`)