fix: effectivize the execution of the 'cs' process
Before this multiple execution of 'cs review' could be in flight for the
same document. But only the latest execution is ever of interest, so it
makes no sense to keep the other ones running. This change refactors the
execution of external processes so that we can keep this under control.