feat: Code Health Review panel, a view of analysed files
Add an optional panel to the explorer view, showing the state of all reviewed/analysed files.

The panel is configured to no be shown by default for now.
Note: Due to some disturbing behaviour the view only shows up after reloading vscode when enabling the setting. After that, toggling it will hide/show it as expected.

There are also some notable changes in the reviewer, handling the fact that we rely on the CachingReviewer implementation for much of the functionality. Also it's rewritten to better accommodate for the async behaviour and handling state for multiple running reviews.