feat(code-health-monitor): Always compare changes to the current HEAD of the file
On opening a file, both the current content and the HEAD version will be analysed. The baseline score will be used in detecting degradations in the code health monitor

Also makes sure to reset the baseline review when HEAD commit changes
This is done by tapping into the vscode.git extension api and listening for state changes in the repositories. The vscode.git extension also listens to extenal changes, so these will be picked up by our extension as well.

Furthermore, we're adapting the baseline review to use the new ide-api + cleaning up json input to delta cmd