fix(code-health-monitor): Invalidate current refactorings for a file when applying a refactoring
This is to prevent accidentally applying the same refactoring twice. Instead it makes sense to invalidate and await a new review+delta analysis of the updated file to get it re-populated with other possible refactoring suggestions.

This commit also moves some of the functionality used by the refactoring panel into the refactoring/commands module.