feat(code-health-monitor): Show fixed issues and improvement opportunities
Using the new ide binary we can get a more fine-grained "delta" analysis. This is reflected in the code-health-monitor by showing fixed issues or improved issues as well as the degradations or introduced issues.
Sorting is done in both the tree view on fn level and in the details view on issue level to get the degradations at the top, followed by improved (but not entirely fixed) issues and fixed issues last.